Type acquisition/threat range globals as backend words
g_acqRange and g_nearestThreatRange carry mapRange results, which are
fractional and uncapped under the modern backend; int16/uint16 stores
re-narrowed them, and the uint16 wrap could turn a >64k modern range
into a small distance. best/dist locals, wordAbs() for the
abs((int16)best) idiom, and dropping the (unsigned) compare casts follow.
g_acqAimY stays a word — word-domain out-param contract.

+++ src/egframe.c
@@ -260,7 +260,7 @@ void updateFrame(void) {
             const auto dist = mapRange(mapOffset(flightMapPosition(),
                                                  g_planeTable.planes[i].mapX, g_planeTable.planes[i].mapY));
             if (dist < g_nearestThreatRange) {
-                g_nearestThreatRange = (int16)dist;
+                g_nearestThreatRange = dist;
                 g_closestThreatIndex = i;
             }
         }
